Incorporate changes from our 2014-ocha-fork.
Some of these changes are directly based on the unit testing in
`wellnest`/`p3p`.  Others are cargo-culted, based on a belief that there
was probably a good reason the XPath changes were made originally.

None of the changes should be significant security-wise, and we should
be able to rebase off `onelogin/ruby-saml` going forward.